Refactor(#30-battery-view): BatteryDashboardSleepView 파일 분리
